pyproject.toml
[project]
name = "qwenpaw"
dynamic = ["version"]
description = "QwenPaw is a **personal assistant** that runs in your own environment. It talks to you over multiple channels (DingTalk, Feishu, QQ, Discord, iMessage, etc.) and runs scheduled tasks according to your configuration. **What it can do is driven by Skills — the possibilities are open-ended.** Built-in skills include cron, PDF/Office handling, news digest, file reading, and more; you can add custom skills. All data and tasks run on your machine; no third-party hosting."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "agentscope==1.0.18",
    "agentscope-runtime==1.1.3",
    "httpx>=0.27.0",
    "packaging>=24.0",
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "alibabacloud-dingtalk>=2.2.42",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2,<4",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "mss>=9.0.0",
    "reme-ai==0.3.1.8",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-socks>=2.5.3",
    "onnxruntime<1.24",
    "lark-oapi>=1.5.3",
    "python-telegram-bot>=20.0",
    "twilio>=9.10.2",
    "pywebview>=4.0",
    "aiofiles>=24.1.0",
    "paho-mqtt>=2.0.0",
    "wecom-aibot-python-sdk==1.0.2",
    "matrix-nio>=0.24.0",
    "shortuuid>=1.0.0",
    "google-genai>=1.67.0",
    "tzdata>=2024.1",
    "cryptography>=43.0.0",
    "keyring>=25.0.0",
    "pyyaml>=6.0",
    "json-repair>=0.30.0",
    "segno>=1.6.6",
    "modelscope>=1.35.0",
    "huggingface_hub>=0.20.0",
    "pillow>=10.0.0",
    "pymysql>=1.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "openpyxl>=3.1.0",
    # anyio 4.13.0: _deliver_cancellation busy-loop / getpid storm (QwenPaw#2632)
    "anyio>=4.0.0,<4.13.0",
]

[tool.setuptools.dynamic]
version = {attr = "qwenpaw.__version__.__version__"}

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
# Dockerfile copies console build output to src/qwenpaw/console/ (no dist subdir).
"qwenpaw" = [
    "console/**",
    "agents/md_files/**",
    "agents/skills/**",
    "tokenizer/**",
    "security/tool_guard/rules/**",
    "security/skill_scanner/rules/**",
    "security/skill_scanner/data/**",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
qwenpaw = "qwenpaw.cli.main:cli"
copaw = "qwenpaw.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
    "hypothesis>=6.0.0",
]
local = [
    "huggingface_hub>=0.20.0",
]
llamacpp = [
    "qwenpaw[local]",
    "llama-cpp-python>=0.3.0",
]
mlx = [
    "qwenpaw[local]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]
ollama = [
    "ollama>=0.6.1",
]
whisper = [
    "openai-whisper>=20231117",
]
full = [
    "qwenpaw[local,ollama,llamacpp,whisper]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
